Feat: useCart, cart
