+++ src/refactoring/hooks/useCart.ts
@@ -1,12 +1,13 @@
 // useCart.ts
 import { useState } from "react";
 import { CartItem, Coupon, Product } from "../../types";
-// import { calculateCartTotal, updateCartItemQuantity } from "../models/cart";
+import { calculateCartTotal, updateCartItemQuantity } from "../models/cart";
 
 export const useCart = () => {
   const [cart, setCart] = useState<CartItem[]>([]);
   const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
 
+  /**장바구니에 상품을 추가합니다. */
   const addToCart = (product: Product) => {
     const remainingStock = getRemainingStock(product);
     if (remainingStock <= 0) return;
@@ -26,78 +27,40 @@ export const useCart = () => {
     });
   };
 
+  /**남은 재고 수량을 반환합니다.*/
   const getRemainingStock = (product: Product) => {
     const cartItem = cart.find((item) => item.product.id === product.id);
     return product.stock - (cartItem?.quantity || 0);
   };
 
+  /**장바구니에서 상품을 삭제합니다 */
   const removeFromCart = (productId: string) => {
     setCart((prevCart) =>
       prevCart.filter((item) => item.product.id !== productId),
     );
   };
 
+  /** 재고수량을 업데이트 */
   const updateQuantity = (productId: string, newQuantity: number) => {
     setCart((prevCart) =>
-      prevCart
-        .map((item) => {
-          if (item.product.id === productId) {
-            const maxQuantity = item.product.stock;
-            const updatedQuantity = Math.max(
-              0,
-              Math.min(newQuantity, maxQuantity),
-            );
-            return updatedQuantity > 0
-              ? { ...item, quantity: updatedQuantity }
-              : null;
-          }
-          return item;
-        })
-        .filter((item): item is CartItem => item !== null),
+      updateCartItemQuantity(prevCart, productId, newQuantity),
     );
   };
 
+  /**쿠폰 적용*/
   const applyCoupon = (coupon: Coupon) => {
     setSelectedCoupon(coupon);
   };
 
+  /**상품금액, 최종 결제 금액, 할인 금액을 반환합니다.  */
   const calculateTotal = () => {
-    let totalBeforeDiscount = 0;
-    let totalAfterDiscount = 0;
-
-    cart.forEach((item) => {
-      const { price } = item.product;
-      const { quantity } = item;
-      totalBeforeDiscount += price * quantity;
-
-      const discount = item.product.discounts.reduce((maxDiscount, d) => {
-        return quantity >= d.quantity && d.rate > maxDiscount
-          ? d.rate
-          : maxDiscount;
-      }, 0);
-
-      totalAfterDiscount += price * quantity * (1 - discount);
-    });
-
-    let totalDiscount = totalBeforeDiscount - totalAfterDiscount;
-
-    // 쿠폰 적용
-    if (selectedCoupon) {
-      if (selectedCoupon.discountType === "amount") {
-        totalAfterDiscount = Math.max(
-          0,
-          totalAfterDiscount - selectedCoupon.discountValue,
-        );
-      } else {
-        totalAfterDiscount *= 1 - selectedCoupon.discountValue / 100;
-      }
-      totalDiscount = totalBeforeDiscount - totalAfterDiscount;
-    }
+    const { totalBeforeDiscount, totalAfterDiscount, totalDiscount } =
+      calculateCartTotal(cart, selectedCoupon);
 
     return {
-      totalBeforeDiscount: Math.round(totalBeforeDiscount),
-      totalAfterDiscount: Math.round(totalAfterDiscount),
-      totalDiscount: Math.round(totalDiscount),
+      totalAfterDiscount,
+      totalBeforeDiscount,
+      totalDiscount,
     };
   };
